Minimizing memory footprint

+++ resources/shaders/withTexture.vert
@@ -1,6 +1,6 @@
 #version 400
 layout (location = 0) in ivec3 aPos;
-layout (location = 1) in vec2 aTexCoord;
+layout (location = 1) in ivec2 aTexCoord;
 
 out vec2 vTexCoords;
 out float vVisibility;
