+++ src/utils/vertex.h
@@ -6,7 +6,7 @@
 
 struct Vertex {
     u8vec3 position;
-    vec2 uv;
+    u16vec2 uv;
 };
 
 #endif //BLOCKGAME_VERTEX_H
